Clean up and fix the projection observation tests
These tests checked that the change notification had either the expected
property value or `nil`. Due to a mix of performing the write on the same
thread as the observation and invalid dynamic casts (e.g. `as? [E]` for a value
which is `List<E>`), the values were *always* `nil` and the tests would pass
regardless of what actual values were sent.